Call docker stop/rm only when there are sigma containers running

+++ stop.sh
@@ -2,10 +2,12 @@
 
 if mount | grep -q 9p; then
     echo "umount /mnt/9p"
-    $DIR/umount.sh
+    ./umount.sh
 fi
 
 pgrep -x proxyd > /dev/null && killall -9 proxyd
 
-docker stop $(docker ps -a | grep 'sigma' | cut -d ' ' -f1)
-docker rm $(docker ps -a | grep 'sigma' | cut -d ' ' -f1)
+if docker ps -a | grep 'sigma'; then
+    docker stop $(docker ps -a | grep 'sigma' | cut -d ' ' -f1)
+    docker rm $(docker ps -a | grep 'sigma' | cut -d ' ' -f1)
+fi
